大改mixin

+++ src/main/java/org/confluence/terra_curio/common/init/TCAttributes.java
@@ -111,10 +111,10 @@ public final class TCAttributes {
         }
     }
 
-    public static float applyArrowKnockback(Entity attacker, float original) {
+    public static double applyArrowKnockback(Entity attacker, double original) {
         if (attacker instanceof LivingEntity living) {
             AttributeInstance instance = living.getAttribute(Attributes.ATTACK_KNOCKBACK);
-            if (instance != null) return (float) (original * (1.0 + instance.getValue()));
+            if (instance != null) return original * (1.0 + instance.getValue());
         }
         return original;
     }
